WIP: add base capi provider
Adds a default CAPI provider to serve as a base implementation,
which cloud providers should embed. By embedding this implementation,
providers can choose to "override" a function from the Provider
interface, but can "skip" implementing a function if the default
behavior is suitable.

Furthermore, embedding the default providers provides a mechanism
to supply a default behavior if the Provider interface needs to grow.